+++ Cython/Compiler/Options.py
@@ -101,6 +101,7 @@ directive_defaults = {
     'profile': False,
     'no_gc_clear': False,
     'linetrace': False,
+    'annotation_typing': False,  # read type declarations from Python function annotations
     'infer_types': None,
     'infer_types.verbose': False,
     'autotestdict': True,
@@ -228,6 +229,7 @@ directive_scopes = { # defaults to available everywhere
     'test_assert_path_exists' : ('function', 'class', 'cclass'),
     'test_fail_if_path_exists' : ('function', 'class', 'cclass'),
     'freelist': ('cclass',),
+    'annotation_typing': ('module',),  # FIXME: analysis currently lacks more specific function scope
     # Avoid scope-specific to/from_py_functions for c_string.
     'c_string_type': ('module',),
     'c_string_encoding': ('module',),